monitor_orders/requirements.txt
fastapi
uvicorn[standard]
PyJWT
passlib[bcrypt]
mysql-connector-python
redis
python-dotenv
kiteconnect
pandas
requests